move _liana_pipe
