+++ liana/method/sc/_natmi.py
@@ -1,4 +1,4 @@
-from liana.method._Method import Method, MethodMeta
+from liana.method.sc._Method import Method, MethodMeta
 
 
 def _spec_weight(ligand_means, ligand_means_sums, receptor_means, receptor_means_sums):
