+++ liana/method/sc/_singlecellsignalr.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from liana.method._Method import Method, MethodMeta
+from liana.method.sc._Method import Method, MethodMeta
 
 
 def _sca_score(x):
