+++ liana/tests/test_get_mean_perms.py
@@ -4,7 +4,7 @@ from scanpy.datasets import pbmc68k_reduced
 from pandas import read_csv
 
 from liana.method._pipe_utils._get_mean_perms import _get_means_perms, _get_positions
-from liana.method._liana_pipe import _trimean
+from liana.method.sc._liana_pipe import _trimean
 
 test_path = pathlib.Path(__file__).parent
